requirements.txt
flask
flask-cors
pandas
langchain
openai
python-dotenv
faiss-cpu
tiktoken
langchain-openai
langchain-community
faker
pypdf